linter: remove nep-1
Summary

Drop nep-1 as the it's an awful standard, replaced with a more lax
implementation that can be tightened up over time.

Details

This remove all mentions of nep-1, updates the style checker to be
closer to symbol kind --> naming style policy map. This policy in turn
is applied to the name string.

The styles remain an incredible mess and there really isn't a great way
to setup this rules based on the current level of information and style
control input. A major overhaul is required in the future, this is a
small stepping stone.

With that said, updated the internals, cleaned up the docs, and added a
few tests. Additionally, moved the `tlinter` test to the `stylecheck`
category along with the rest of the style related tests.