parser: remove unused `pnkImportAs` node kind
Summary

This node kind is not produced by the parser and is handled entirely in
sem, as such it was removed. No functional change.

Details

Removed the kind and appropriately updated conversion parsed to ast node
conversion routines to account for this.

This error was potentially introduced because not all parser node kinds
are explicitly mentioned in the `parser` module. The import and export
kinds in particular used `succ` to compute the except conditions. This
requires the human to "play computer" and result in mistakes such as
these. The code now explicitly maps types without relying on implicit
ordering.