make `vmbackend` compile again
The `ModuleList` used by the collector pass is now no longer re-used in
the VM backend, but instead translated to a custom representation (see
`produceModules`).

In addition, the list with the top-level AST is now discarded after it's
passed to code generation, which should reduce memory pressure, and thus
peak memory usage, quite a bit.