collectors: track a module's `IdGenerator`
The upcoming C code-generation orchestrator needs access to the
`IdGenerator` for each module. They're now stored with `FullModule`,
allowing the orchestrator to retrieve them later.

`vmbackend` is adjusted to make use of the module's `IdGenerator`, which
is more precise than using the `ModuleGraph`'s one. Ideally, the backend
should not introduce any new symbol and type instances (which is what
the `IdGenerator`s are required for), but `transf` currently
necessitates that.