add a C code-generation orchestrator
Much like how it works for the VM backend, the C backend now also uses
an orchestrator (the new `cbackend` module) that invokes the code
generator.

The `passes` integration (`myOpen`, `myProcess`, and `myClose`) is
removed from `cgen`. Setting up the extra header backend module (used
for C header generation) is now the responsibility of the orchestrator.

Similar to `vmbackend`, the new `cbackend` also operates on the
semantically analyzed AST of the whole program.