collectors: simpler module list management
Store the modules in a `SeqMap` with their position as the key. This
conveys intention better and leaves it to the consumer to use a more
fitting structure/representation.

For the internal representation, `vmbackend` uses a `Store` for the
list of modules, which makes the intention more clear and querying the
list a bit more ergonomic.