jsgen: remove `inSystem`
Disabling stack-traces for the system module and all modules it imports
differs from what the C code-generator does (stack-traces are only
disabled for the `system` module there, but not for the ones it
imports).

The required access to `PGlobals` via `g.backend` is also a small
problem for upcoming `jsgen` refactoring.