parser: graceful keyword in ident position error
Summary

The parser now makes progress even if an keyword is present in an
identifier position.

Details

Previously the `parser` module did not gracefully handle keywords
present in identifier positions, such as `add` in `var add = 10`, here
`add` needs to be escaped, via accent quotes. Instead it stopped
progressing and this led to spurious invalid indentation errors and
early termination of incremental parsing.

Now when a keyword is found in an identifier position, the keyword token
 is turned into an identifier node and processed as if it was
 successful. This means a diagnostic is generated for the error, but the
 token is still consumed (progress).

 After the change the `07_keywords_fail` language specification test now
 works correctly.

 As part of this change an opportunistic refactoring of the
 `lexer.getEscapedChar` was conducted to clean-up some code.