vmbackend: basic support for lifetime hooks
Summary
=======

Enable the `injectdestructors` pass for the VM backend, which is the
first step towards full lifetime hook support there. They currently have
the same limitations as they do with the JavaScript backend, with the
additional one that destructors do not work for top-level globals.

In addition, the `wasMoved` magic procedure is now supported for code
running in the VM (which includes code running at compile time).

This change is meant as a preparation for unifying the backend
processing.

Details
=======

* use `canonicalizeWithInject` (which runs the `injectdestructors` pass)
  instead of just `canonicalize` in `vmbackend`
* add support for the `mWasMoved` magic for `vmgen` by re-using the
  `mReset` implementation
* enable tests from the `lang_objects/destructor` category for the VM
  target where doing so is possible
* enable the `tforloop_tuple_unpacking.nim` test for the VM target

The limitations of lifetime hooks for the VM backend currently are:
* destructors aren't called for refcounted heap cells (i.e., `ref`s)
* destructors aren't called for top-level globals
* the copy and destroy hooks are not called for `seq` elements when the
  `seq` is copied or destroyed
* cursors are ignored and no shallow copies are used when assigning to
  them