spec: introduced bootstrap specification
Summary

Introduced a "boostrap" specification that is meant to test a barebones
compiler implementation. This new arrangement of code is sets the
revised tone of the the specification.

Details

This mostly rearranges the tests. While working on these particular
specs it was found that the identifier related lexer errors were
unclear. These have been revised and their generation is entirely driven
by the `lexer` module itself. `cli_reporter` now simply defers to it. As
part of the change `LexerReport` have lost a lot of their bloat.