Saem specs update test organization (#713)
## Summary

Introduced a "bootstrap" specification that is meant to test a barebones
compiler implementation. This new arrangement of code sets the
revised tone of the specification.

## Details

This mostly rearranges the tests. While working on these particular
specs it was found that the identifier related lexer errors were
unclear. These have been revised and their generation is entirely driven
by the `lexer` module itself. `cli_reporter` now simply defers to it. As
part of the change `LexerReport` have lost a lot of their bloat.

Finally, removed testament's duplicate spec detection.
This only gets in the way of declaring a number of string
literal tests.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>