std/parseopt: only parses command arg vectors (#722)
## Summary

The `parseopt` module no longer supports parsing arbitrary strings as
command line parameters. Instead a `seq[string]` representing an argv
(argument vector) must be supplied. In addition to removing the API for
this, it no longer automatically fetches command line parameters.

In addition, parsing empty argument lists and short flags with value was
fixed. In particular, input such as `["-s", "foo"]`, where `"-s"`
expects an argument is correctly treated as `-s:foo` or `-s=foo`.

## Details

As part of this change the `string` accepting variant of the
`initOptParser` procedure was removed, along with the similarly
misguided
`getOpts` iterator. Lastly, an empty list no longer triggers argument
retrieval via the `os` module, meaning these are handled correctly.

Modified the compiler and various executables to transition to the
revised API, in order to facilitate this introduced the `os.getExecArgs`
procedure.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>
Co-authored-by: Saem Ghani <saemghani+github@gmail.com>