injectdestructor: adjust to the global variable changes
Checking for threadvars and globals defined in a module's outermost scope
is no longer necessary, as their definition statements are removed by the
MIR pass running prior to the `injectdestructors` pass.

We only need to check for pure globals now, but once those are properly
extracted when using the JS backend, the last-remaining special-casing
for globals can be removed.