implement code generation for module destructors
The module de-initialization procedures (also refered to as module
destructors) are a bit tricky to handle. We only know their contents
*after* we know all alive globals, but thanks to pure globals, we only
know all alive globals after all alive procedures were discovered and
processed.

Mirroring the original logic, we finish the module destructors right
after wrapping up the main part of code generation for the main module.
Same as with the original logic, this means that new globals discovered
past that point will not be considered for destruction (or
initialization).