backend: simpler error handling (#736)
## Summary

Consistently disable the code generation phase across backends when
there were errors during the semantic phase, and make sure that the
first error reported during code generation terminates the compiler.

This prevents errors destabilizing the code generators when
`errorMax` > 1.

## Details

The goal is to eventually remove all raising of semantic errors from the
code generators, so overriding the user-provided error upper-bound in
the meantime is acceptable.