compilerbridge: make `PEvalContext` a standalone object
The `ModuleGraph` now stores an instance of `PEvalContext`, which is
going to allow for splitting the `TCtx` type up into its constituent
parts. To get around a circular dependency involving `sealRodFile`, the
procedure is moved to the `sem` module, which is also the only place
where it's used.

So far, only the `oldErrorCount` counter is moved to the
`PEvalContext`.