IC: make the IC backend work again
For simplicity, the IC backend now acts as a front-end for the C code-
generation orchestrator, being responsible for turning the rodfile
contents into a `ModuleList`.

While this is going to make further changes to `cgen` much easier, as
the IC backend no longer has to be special cased as much, it has the
significant downside of always having to run code generation again,
even for modules that did not change.