cgen: fix `reset` and `default` for resolved concepts
Summary
=======

Skip user type-classes when generating code for `reset` and `default`,
fixing semantically invalid C code being generated when the operand is
a resolved user-type-class.

Details
=======

When not skipped, the branch for emitting assigning `0` was chosen.
Skipping the type in `cgen` is fundamentally a workaround, resolved
user-type-classes should not reach the code generators in the first
place.