codegen: clean up unused code paths (#740)
## Summary

Remove unused code paths in `cgen`, `jsgen`, and `vmgen`.

Besides the general benefits that removing unused code paths has, this
also helps with preventing regressions caused by lowerings not being
properly applied, at least until the the code generators are rewritten.

## Details

The AST shapes for which the code paths are removed are:
- `when nimvm` statements
- `if`, `case`, `block`, and `try` *expressions*
- multi-branch `if` statements
- `while x` statements where `x` is something else than a `true` boolean
  literal
- pragma blocks and lambda-expressions
- enum field symbols
- `and`/`or` expressions
- deconstruction in `var`/`let` statements (i.e. `nkVarTuple`) and
  initialization of lifted locals (e.g., `var env.x = 0`)

Except for `skEnumField` symbols (which are lowered away during
`transf`), all of the above are lowered during the MIR phase
already, and never reach code generation.

There are two small changes to the output of the code generators:
`jsgen` no longer emits a redundant `if (!true) break label;` statement
for `while` loops, and `cgen` doesn't lower `if` into gotos anymore. An
example of the latter:
```c
// `if cond: then` now produces: 
if (cond)
{
  then;
}
// instead of:
if (!cond) goto label;
then;
label:
```