cgen: don't emit unnecessary case statements (#739)
## Summary

Fix a logic bug in `cgen`'s object-init specialization that led to
potentially large but empty C case statements being emitted.

While the bug had no effect on program semantics, it did cause more C
code to be emitted, which translated to more work for both the NimSkull
and C compiler.

## Details

The predicate was the inverse of `isObjLackingTypeField`, but that is
incorrect. `isObjLackingTypeField` also returns false for anything
besides `tyObject`, so in effect, all objects that had fields of
non-object types were erroneously treated as having a type field
somewhere.

Also adjust stale documentation related to object-init specialization.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>