mirgen: improve lowering of `and`/`or`
Summary
=======

Don't use a `block` + `break` for lowered `and`/`or` expressions. This
results in less MIR code and smaller control-flow graphs (only a single
`fork` and `join`, instead of 1 `fork`, 1 `goto`, and 2 `join`s).

In addition, all code generators produce less code for the new lowered
version than they did for the previous one.

Details
=======

The MIR code that is now generated by `mirgen` is, except for the
scoping behaviour, equivalent to:
```nim
  # input expression: `a and b`
  var tmp
  tmp = a
  if tmp:
    tmp = b
```
Previously, it looked like:
```nim
  var tmp
  block label:
    tmp = a
    if not tmp: break label
    tmp = b
```