cgen: fix `.global`s not being initialized
A bug local to this PR. If the initializer is sufficiently simple, the
logic in `genSingleVar` still expected `assignGlobalVar` to make sure
that the global was initialized. This isn't the case anymore, so the
globals were left uninitialized.

In addition to removing `potentialValueInit`, `assignGlobalVar` is also
removed, as it does nothing else besides dispatching to
`definedGlobalVar`.