mirbridge: fix initialization of globals
Just removing the `def` in case it doesn't have an initializer is wrong.
Globals defined in loops and globals for which the zero-representation
is not valid (e.g. those that have type fields) need to be explicitly
set to their default value.

In order to also support modules being set-up and destroyed multiple
times during the lifetime of a program (not possible right now, but
might be in the future), assigning the default value is performed for
all globals.

There were also two other bugs with the rewriting logic:
- `mnkDefUnpack` and `mnkDefCursor` weren't considered
- the `mnkInit` source argument use the `arg` mode instead of the
  `consume` mode