consistent canonical import and package sym naming (#746)
## Summary

Package symbols for modules are now generated in a consistent fashion:

- at the root of a package: `pkgroot/module`
- nested within a package: `pkgroot/pkgpath/module`
- modules within the project package omit `pkgroot/`

This leads to more consistent messages and a variety of improvements for
internals (eg: more stable signature hashes) and a better foundation to
build upon.

## Details

Previously, package symbols were created and modules added to them in a
flat manner, until there was a name collision (two modules in different
directories with the same name). In the case of collisions a mangled
package name was created based on the module that caused the collision
and that was used instead. Although the package symbol information isn't
used in too many parts of the compiler this led to yet more broken data
and subtle issues for anything that relied upon it:

- signature hashing owners and procedures
- c, rod, and various file name generation
- error messages: canonical imports
- and more

Getting package info is now done via `packagehandling.getPkgDesc`, which
retrieves various key details, including dealing with the lack of
package definition. Package detection still remains coupled to nimble
files, for now.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>