compiler: partially remove `codegenDecl` pragma (#748)
## Summary

Altering the code generators' output through programmable string-
formatting (which is what the pragma provides) does not align with
NimSkull's goals, and the feature is generally becoming more and more
unstable as a side-effect of the backend-processing unification.

It is now mentioned as deprecated in the manual, only allowed on
`proc`s, and treated as having no effect when not using the JavaScript
backend.

## Details

The `asyncjs` module is the only remaining user of the pragma, and since
the VSCode extension currently depends on said module, the pragma is not
fully removed yet.

* remove all `.codegenDecl` support from `cgen`
* disallow the pragma on `var`/`let` statements
* remove all tests related to the pragma
* adjust the manual