sequtils: simplify `mapIt` implementation
Summary
=======

Use an inline iterator instead of a macro for achieving the "evaluate
once" semantics in the `mapIt` template. This makes the code simpler,
and, given that the compiler's code itself uses `mapIt`, also reduces
the amount of macro usage during bootstrapping.

Details
=======

* introduce the internal `mapIter` iterator
* remove the now obsolete internal `evalOnceAs` macro
* adjust the internal documentation of `mapIt` a bit
* adjust the `t8339.nim` test to not depend on `sequtils`
* don't introduce an unnecessary shallow copy when lifting `op` into a
  procedure

While the `mapIter` doesn't use `lent`, special care is taken for it to
be eligible for the cursor optimization. With the current compiler, for-
vars are always turned into cursors when the invoked inline iterator
only ever yields the *symbol* of a local.

The `t8339.nim` test is intended to make sure that aliases are skipped
when rendering types, but using `sequtils` for testing that is not
required, and only made the test unnecessarily depend on the `mapIt`
implementation details (which now changed, making the test fail).