std/parseopt2: new exec argument parsing library (#729)
## Summary

Brand new and simpler implementation of 'parseopt', the new module does
not replace `parseopt` to allow the compiler code base time to migrate.

## Details

The parsing approach takes after posix getOpt, taking on conventions
such as space separated flag arguments and treating all args after `--`
as positional.

The new API is main the `opts` iterator which parse input and produces
`Opt` for each option, and input errors are inline via a field.

Along with the new module is a comprehensive set of tests.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>