jsonutils: no unmarshal support for `cstring` tables
Summary
=======

The direct unmarshal (`fromJson`) approach used by the `fromJsonHook`
implementation for `Table`s or `OrderedTable`s where the key is a
`cstring` cannot work (at least when using the C backend), and the
overload is thus removed.

Details
=======

The reason that it cannot work for the C target, is that there,
`cstring`s are unsafe views into strings, so the keys of the
unmarshalled table are only valid (at maximum) for as long as the source
`JsonNode` is alive (and unchanged).

Due the iterators used by `fromJsonHook` not using `lent` and thus
introducing temporary copies, the `cstring` keys were only valid for
the loop iteration in which they were added to the table.

The test case that existed for `OrderedTable[cstring, int]` only worked
by accident. Changing the allocation behaviour, building with
`-d:useMalloc`, or enabling memory sanitizers would all result in
test failures.