jsonutils: no unmarshal support for `cstring` tables (#752)
## Summary

Unmarshaling a `Table` or `OrderedTable` where the key is a `cstring`
via `fromJsonHook` cannot be implemented to work in a safe way, at
least when using the C backend. For this reason, the overload is
removed.

## Details

The reason that it cannot work for the C target, is that `cstring`s
are unsafe views into strings there, so the keys of the table can,
in theory, only be valid for as long as the source `JsonNode` is
alive (and unchanged).

In practice, due to the iterators used by `fromJsonHook` not using
`lent` and thus introducing temporary copies, the `cstring` keys were
only valid for the loop iteration in which they were added to the
table.

The test case that existed for `OrderedTable[cstring, int]` only worked
by accident. Changing the allocation behaviour, building with
`-d:useMalloc`, or enabling memory sanitizers would all result in
test failures.