vmgen: fix accessing `var`/`lent` views in objects
Summary
=======

Make `vmgen` emit the bytecode for loading pointer views into a register
prior to the dereference, making access of `var`/`lent` views stored in
objects work for code being compiled for the VM target (refer to the
added tests for examples; most of them previously crashed the VM).

Details
=======

The VM can only dereference pointers - which direct single-location
views part of compound types map to for the VM target - if they're
stored in registers. Due to the `vmgen` logic for generating the code
for `nkDotExpr` (field access) and `nkBracketExpr` (array and seq
access) expressions not considering views when testing whether a load-
into-register is required, a handle-to-pointer operand was passed to
`LdDeref`.

`needsRegLoad` now checks whether a load-into-register is required via
the new `fitsRegisterConsiderView`, which treats direct single-location
views as fitting into registers.

Future Direction
----------------

The code generator needs to map types to a register / internal type and
then base the various type queries it performs (`isDirectView`,
`fitsRegister`, etc.) on the mapped type.