mirgen: remove `nkHiddenAddr`/`nkHiddenDeref` for `openArray`
Semantic analysis inserts `nkHiddenAddr` and `nkHiddenDeref` for
creation and access of `var|lent openArray` values. This is a problem
for the MIR and the code generators, as a `var|lent openArray` is (for
the largest part) the same as an `openArray` there.

`mirgen` now removes the unnecessary operations; the small formal
vs. actual type mismatch that results is acceptable for the time being.

This fixes assigning `seq`s, `array`s, etc. to `var openArray`
locations.