refactor(ast): minimize usage of `nfBlockArg` flag (#755)
## Summary

Only statement list nodes are flagged with `nfBlockArg` as a purely
internal refactoring.

## Details

The `nfBlockArg` flag is used to indicate that a statement list
(`nkStmtList` or `nkStmtListExpr`) is a block argument and not to be
unwrapped if containing a single expression. The flagging is done in the
`parser` and was being applied to non-statement list nodes where such
guards are not required.

Also removed a unnecessary unwrapping guard in `semStmtList`.

Guarding against unwrapping is an operand handling matter and should
probably be handled as part of `sigmatch`, in `prepareOperand` and
`semOperand`.